requirements.txt
line-bot-sdk
flask
selenium
webdriver-manager
beautifulsoup4
python-dotenv